DSND-2879: WIP Embedded Kafka tests now work with test containers
