+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 server.port=8081
 springfox.documentation.auto-startup=false
-spring.kafka.bootstrap-servers=${STREAMING_KAFKA_BROKER_URL:localhost:9092}
+spring.kafka.bootstrap-servers=${BOOTSTRAP_SERVER_URL:localhost:9092}
 spring.kafka.listener.concurrency=${COMPANY_METRICS_LISTENER_CONCURRENCY:1}
 company-metrics.consumer.charges.stream.enable=${COMPANY_METRICS_CONSUMER_CHARGES_ENABLED:true}
 company-metrics.consumer.charges.stream.group-id=${COMPANY_METRICS_CONSUMER_CHARGES_GROUP_ID:company-metrics-consumer}
