--- src/test/resources/application-test_consumer_main.properties
@@ -1,20 +1,0 @@
-server.port=8080
-spring.kafka.listener.concurrency=1
-company-metrics.consumer.charges.stream.enable=true
-company-metrics.consumer.charges.stream.group-id=company-metrics-consumer
-company-metrics.consumer.charges.stream.topic=stream-company-charges
-company-metrics.consumer.charges.stream.retry-attempts=4
-company-metrics.consumer.charges.stream.backoff-delay=100
-company-metrics.consumer.appointments.stream.enable=true
-company-metrics.consumer.appointments.stream.group-id=company-metrics-consumer
-company-metrics.consumer.appointments.stream.topic=stream-company-officers
-company-metrics.consumer.appointments.stream.retry-attempts=4
-company-metrics.consumer.appointments.stream.backoff-delay=100
-company-metrics.consumer.psc-statements.stream.enable=true
-company-metrics.consumer.psc-statements.stream.group-id=company-metrics-consumer
-company-metrics.consumer.psc-statements.stream.topic=stream-psc-statements
-company-metrics.consumer.psc-statements.stream.retry-attempts=4
-company-metrics.consumer.psc-statements.stream.backoff-delay=100
-logger.namespace=company-metrics-consumer
-api.api-url=http://localhost:8888
-api.company-metrics-api-key=${COMPANY_METRICS_API_KEY:localhost}
